mlops_cluster/cluster/iris/app/requirements.txt
sqlalchemy
python-dotenv
pandas
scikit-learn
mlflow
psycopg2-binary